packages/mmos-squad/requirements.txt
# MMOS Expansion Pack Dependencies

# Core dependencies
pyyaml>=6.0
requests>=2.31.0
google-generativeai>=0.3.0  # Gemini API for cost-efficient materials analysis

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Database
# sqlalchemy>=2.0.0
